+++ cpp/src/dmrpt/io/image_reader.hpp
@@ -19,9 +19,9 @@ class ImageReader {
   read_mnist_labels (string path, int no_of_images, int dimension, int rank, int world_size);
   vector <vector<VALUE_TYPE>> read_File (string path, int no_of_data_points, int dimension, int rank, int world_size);
   vector <vector<VALUE_TYPE>>
-  mpi_file_read (string path, int rank, int world_size, int overlap, int total_data_set_size, char delim, int dimension);
+  mpi_file_read (string path, int rank, int world_size, int overlap, long total_data_set_size, char delim, int dimension);
   vector <vector<VALUE_TYPE>>
-  mpi_file_read (string path, int rank, int world_size, int overlap, int total_data_set_size, int data_node_byte, int offset, int dimension);
+  mpi_file_read (string path, int rank, int world_size, int overlap, long total_data_set_size, int data_node_byte, int offset, int dimension);
   int reverse_int (int i);
 };
 }
